<SWAUTO-3464> Add device description properties to the DataSink Capture FB

+++ asam_cmp_capture_module/tests/test_stream.cpp
@@ -18,7 +18,7 @@
 using namespace daq;
 using namespace testing;
 
-class StreamFbTest: public testing::Test
+class StreamFbTest : public testing::Test
 {
 protected:
     StreamFbTest()
@@ -71,7 +71,6 @@ protected:
         for (const auto& e : decoder.decode(data.data(), data.size()))
             receivedPackets.push(e);
     };
-
 
     void rawCanFrameCapture(const CANData& data, bool allowCanFd)
     {
@@ -134,7 +133,7 @@ void StreamFbTest::testCanPacketWithParameter(bool isCanFd)
     createProc();
     auto streamFb = interfaceFb.getFunctionBlocks().getItemAt(0);
 
-    uint8_t streamId = streamFb.getPropertyValue("StreamId");
+    uint8_t streamId = static_cast<Int>(streamFb.getPropertyValue("StreamId"));
     uint32_t interfaceId = interfaceFb.getPropertyValue("InterfaceId");
     uint16_t deviceId = captureFb.getPropertyValue("DeviceId");
